Now also showing popup when user intends to leave the course.
Closes https://github.com/corona-school/project-user/issues/224
